refactor: move report code
